[opaque_test] Add dependency on component_manager.cmx
OpaqueTest does not have a dependency on
//src/sys/component_manager:component-manager.

This means that it is possible to build a test and
have the test fail because the component_manager.cmx
component could not be found.

Also remove the unnecessary direct dependency on
component manager for blobfs stress tests.

This happened in fxbug.dev/63817

Bug: 63817
Change-Id: I64b8ee9b9fe589b1ceb93d923c06a65a66f4d231
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449616
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>