[session] Remove assert on exposed services from component_manager_sfw
This check needs to be updated if the session_manager cml manifest
changes (e.g. when exposing a new protocol), and couples
component_manager_sfw to specific protocols. It also runs on each
invocation, not just in tests.

It might be better implemented as an integration test in the future.

Change-Id: Ie6f8e85c5d5364748fe8ac6599b215c00cb97851
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449380
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>