[fdt] Update fdevtools.sh to handle moving our linux binary
We previously did the work of making fdevtools.sh handle the new binary name
(fuchsia_devtools instead of system_monitor) however because our code
was still building from the system_monitor directory in google3, the
directory structure in our prebuilt still has that name. The final piece
of that story is about to land in google3 where we are doing our final
move of the code. We need to handle both structures so we can push that
forward and then we can finally come back and cleanup the _OLD stuff.

As we have done the binary name transition already we can replace the
current _OLD stuff with the current not old stuff as that is no longer
needed.

Change-Id: I01aaea37e7715d656b6256d7d403755f0eb51a26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449171
Reviewed-by: Alex Bogart <abbogart@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>