[fuchsia-runtime] Panic if runtime handle is not available
zx_clock_read will not fail even if the clock has never been updated,
it instead returns the backstop if it was created with one, or 0
otherwise. This means that utc_time() cannot fail unless the process
was launched with no UTC clock or a handle missing the required rights.

The documentation for zx_clock_read is updated in fxr/448994.

Bug: 61741

Change-Id: Ia733fa55ba9a03dfd7bde9447b6bd1ea07f9ea9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449074
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>