[settings] Refactor media buttons agent to use a hashset of settings
This CL updates the media buttons agent to calculate the settings that
it wants to send events to and store it as a set, instead of using an if
statement for each setting. This is done in preparation for adding light
controller as another setting type to receive media button events and
makes it easier to add additional setting types in the future.

Fixed: 64142
Test: fx test setui_service_tests
MULTIPLY: setui_service_tests
Change-Id: I6c7734b0cf9951f0e58889cff3a7c02cee00504a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449717
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>