[testrunner] Don't fail for snapshot failures.
Usually a failure to get a snapshot is really an SSH failure or some
other test failure that caused the snapshot to fail as well.
https://fuchsia-review.googlesource.com/c/fuchsia/+/446298 took care of
returning the SSH failure over the snapshot failure but it doesn't take
care of the case where the failure is caused by a test failure. Since
the snapshot is used more for debugging purposes, we don't need to fail
the task if it fails to run. This way we can surface the actual failure
if it's caused by a test failure. In the case that all tests pass,
however, tefmocheck should find and return the snapshot error.

Bug: 61436
Change-Id: Id8b7b9bbf6fa799f8ce175aeef8a77b7e377ea66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449799
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>