[timekeeper] Address cobalt unittest flake.
There are two potentially causes of a mismatch in the expected clock
difference, being prempted between the two calls to initialize the
test clock objects and being prempted while reading the clock values.

The latter could be a problem leading to inaccurate data in real life
(on an emulator the test is flaking with slightly over 1ms difference)
so this CL has a few attempts to get a close pair (if we were prempted
on the first attempt the second atempt should be right at the start of
the next timeslice so I wouldn't expect more than one failure in a
row). I've also increased the tolerance on the check to help handle a
premption when setting the test clock.

Multiply: timekeeper-tests
Bug: 63184
Change-Id: Id4080a7a82e65c3d3f5e8befe8aa814ca963ea79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449534
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>