[ffx][daemon] Drop targets from `ffx target list`.
When targets are deemed "disconnected" they will now be dropped from
`ffx target list`

This is unit tested but was also tested manually against emulator
targets.

This adds:
-- An atomic state updater for the target that updates the
   connection state.
-- Events for state transitions (currently only used in tests but could be used
   in a target event handler of some kind as well).
-- An MDNS monitoring task that drops a target when the state is left in
   MDNS with no state change and no repsponses (plus a grace period).

The intent of the connection states are that RCS and Fastboot supercede
mDNS, as they are able to more reliably determine if the target is up or
not, and as such will transition directly to
ConnectionState::Disconnected.

Bug: 62784

Change-Id: I55eeb1ab1e722cb2c8a79f638d8651b9d9a03314
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446856
Testability-Review: Andrew Davies <awdavies@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>