[ota-dependency-checker] create ota-dependency-checker component
This change creates the v2 ota-dependency-checker component. In order to
save space, it shares a binary with the pkg-resolver.

The integration testing is done in v1 because we are still in the
process of figuring out how we want to do v2 integration tests. We'll
migrate the tests to v2 in fxbug.dev/64125.

Test: fx test -o ota-dependency-checker-tests
Test: fx test -o ota-dependency-checker-integration-tests
Test: fx test -o grand-swd-binary-tests
Fixed: 63641
Change-Id: I48c0fe3b159a5369a75f342b4f9184ee11ae3a25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449417
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Testability-Review: Ben Keller <galbanum@google.com>