[storage][blobfs]: Add ExportBlobs command.
Operates over a generate blobfs images, taking advantage of |VisitBlobs|
writes each blob as an independant file in an output dir.

Added test to verify behavaior. Previous CL that introduced VisitBlobs
verifies that the contents provided to the visitor match the source, and
here we verify that the exported blobs match the contents of the files
used to generate the image on a first place.

Bug: 58653

TEST=blobfs-host, CQ

Change-Id: I901f64c24cd9bed58272d667a7d5a74fe87f8bca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449137
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>