[digest] Fix bug with padding compact trees
If a HashLists's list was a multiple of the node size then a debug
assert in Digest::Update would fail when trying to pad the node because
Finish had already been called on the Digest.  Allowing PadWithZeros to
called on a finished NodeDigest fixes this.

Bug: 36663
Change-Id: Ifbd4dd1a316440b68a22ed98e93678e4539f73e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445469
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>