[nix] implement edge-triggered polling in epoll
This enables the EPOLLET flag in epoll_ctl. When ORed with
the op flag this will cause epoll_wait to return a file
descriptor only when the file descriptor transitions from
inactive to active.

When a file descriptor controlled by EPOLLET is returned
from epoll_wait, non-blocking I/O should be performed on
the descriptor until EWOULDBLOCK is returned before calling
epoll_wait again. If epoll_wait is called when the file
descriptor is still ready, it may not return the descriptor.

Note that, if I/O is performed until EWOULDBLOCK is returned,
but the file descriptor becomes ready before the next call to
epoll_wait, the descriptor is guaranteed to be returned
appropriately, because of the use of ZX_WAIT_ASYNC_EDGE in
zx_object_wait_async.

Test: fx test nix-unittest-package
Bug: 45709
Change-Id: I4fee5736e41fa31eb9bf37ac367aa92577fb4f54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438656
Commit-Queue: Steve Austin <steveaustin@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Testability-Review: Adam Barth <abarth@google.com>