[zircon] Rename "__USED" with "__ALWAYS_EMIT".
The compiler attribute "__USED" indicates that a particular variable is
used and should be emitted, even if the compiler can't see any
references to it.

It is rare to need the attribute, but it can be easily confused for the
attribute "__UNUSED" (which indicates "don't warn if this variable or
function is unused").

Rename "__USED" to "__ALWAYS_EMIT" to reduce confusion.

Change-Id: I972e80e631386d279b471a2c63ea8960917dd7ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449920
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>