[fvm] Support partition limits.
Add support for limiting a FVM partition size. A maximum byte size can
be set for each partition inside FVM, and slice allocations beyond this
size will fail if it would cause the partition to be larger than that.

The maximum partition size is not persisted. It must be set each time
after boot. We plan to do this from fshsot according to some
configuration values. It defaults to "no maximum" for each partition
until explicitly set.

safemath is used in new code and in some existing place which clarifies
the overflow checking. Checked casts are used in a few loops to make it
clear it's guaranteed the value doesn't overflow (I don't think there
was an actual bug here) and the extend and the count is 0-checked in the
allocate and free functions (this probably isn't a bug either but the
allocation unwind code in the case WriteFvmLocked() relied on some
wraparound and casting details in the 0-case that looks suspicious).

Bug: 64055
Change-Id: I6b197f9b2d7a4417625464e208783132b193a2f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449954
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>