[ramdisk] Fix ramdisk sleep race.
There was a race where the sleep-after block count could be set just as
we were processing a write and it would cause the counts to underflow
which was the cause of a flaky test.

Also fixed a blobfs test which was making the wrong assumption about the
device block size.

Fixed: 63846
Change-Id: I9469a2f92c3ab53e79b81bb7a5baf24650cf85f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450318
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>