[zircon][tests] Provide package container for kernel-clocks test
The kernel-clocks unit test is configured to run in a bringup
configuration as a ZBI item. This test can also run in non-bringup
configurations.  This adds a unittest_package() rule for the
kernel-clocks unit test to allow it to also run in these configurations.

Change-Id: Ic597fb839cc5e4a2368ffb4a0cc477ff812f21b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449831
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: James Robinson <jamesr@google.com>