[kernel][vm] Cache attribution counts in VmMapping
Cache page attribution counts for the VMO range that is mapped by a
VmMapping, and return the cached value if the VMO and the mapping have
not changed.

Introduces a mapping generation count to track modifications to the
VmMapping which alter the VMO range that is mapped. The caching logic
uses the mapping generation count and the VMO's hierarchy generation
count to infer if an update to the cached page count is required.

The mapping generation count is updated from VmMapping::ProtectLocked()
and VmMapping::UnmapLocked().

Test: kernel-unittests

Bug: 56779
Change-Id: I2180c4a5766efa110c20f0172c24b8e6abbb4ba9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449555
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>