[netstack] Inspect per-socket info of closing TCP endpoints.
Currently, TCP endpoints are deleted from the internal endpoints map
on a socket close(). This is premature for TCP as the endpoints can
linger around for sometime after a socket close(). The closing
state-machine has multiple states where it could be waiting for a
timeout or an ACK from the peer. For debug purposes, it is useful to be
able to inspect the endpoint info and stats in such states.

This change updates the endpoints map only on receiving EventHUp from
gVisor stack. This also changes map to have a separate monotonic
counter as the key, thus avoiding usage of local handle value as key which
would not be relevant after closing the handle in close() now.

Fixed: 62372
Test: fx test -v netstack_test -- -test.run=TestTCPEndpointMapClosing -test.count=1000
      fx test -v netstack_test -- -test.run=TestEndpointsMapKey
Change-Id: I149f57df6d0ab6d87fb9ec734400828c487bbb45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443875
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>