[build] Fix ubsan variant build.
Trying to build a Fuchsia package that contains a binary built with the
"ubsan" variant (but not "asan-ubsan") did fail with an error stating
that libclang_rt.ubsan_standalone.so was missing.

The reason for this was that the toolchain runtime dependencies in this
case were not computed properly. This CL fixes the issue.

- build/config/BUILDCONFIG.gn, build/toolchain/variant_tags.gni: Add
  a new "ubsan" variant tag and apply it to any variant that actually
  performs undefined behaviour sanitizing (even "asan-ubsan" et al).

- build/toolchain/runtime/BUILD.gn: Fix a typo and ensure a proper
  compiler flags match when only when "ubsan", but not "asan",is
  present in the current variant's tags list.

  That's necessary because clang's runtime.json doesn't have any
  entry for "-fsanitize=address -fsanitize=undefined", and the
  'address' entry is sufficient for asan-ubsan to work.

For testing, use the following in args.gn:

  select_variant = [
    "ubsan/cat",
    "asan"
  ]

Then do "fx build obj/third_party/sbase/cat/package_manifest.json"

Before this CL, this failed to build. After the CL, this works correctly
and looking at the generated manifest file shows it lists the right
runtime libraries under lib/ubsan/

Bug: None
Testing: See above.
Change-Id: I23670f02adb82276b5e839e584ef9499f171d4a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449655
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>