[component_manager] Fix elf_runner tests
In the elf runner tests, an Echo message was received but the
responder oneshot channel was not unblocked. In the right scenario
the client would see that the oneshot was canceled and would crash
crash the test.

Fixed by resuming the client before ending the test.

Test: fx test elf-runner-test
Bug: 61970
Multiply: elf-runner-test: 100
Change-Id: Idd0995af0b50a0fba31c460430dcf48cf93e9e7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451114
Reviewed-by: Fady Samuel <fsamuel@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>