[nix] Add test case for close()ing an FD in an epoll set
This adds a test case for our epoll emulation that considers the case
of closing a file descriptor while it is in an epoll interest set. This
shows a difference in our current implementation compared to the Linux
behavior.

Bug: 64296
Change-Id: I2728a0932f499005c70788f3bb9541311c2126d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450305
Reviewed-by: Steve Austin <steveaustin@google.com>
Testability-Review: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>