[blobfs] Specify the blob layout at build time
The Blob layout format currently defaults to "padded" so this is a
no-op.  Adding this flag makes it easy to switch between the "padded"
and "compact" layouts while we gain confidence in the new "compact"
layout.

Adding --args='blob_layout_format="compact"' to fx set will switch to
using the compact layout for blobfs.

Bug: 36663
Change-Id: I47b6fcdb9e589c29992e3cff420157d8e4a9dfa2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445174
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>