[host-target-testing] Update to use untarred build artifacts.
Current builds all upload the necessary untarred artifacts to the
artifacts bucket which can be used for paving. As part of the effort to
remove archives, this change will switch the OTA tests to fetch the
untarred artifacts first, and if the required artifacts are missing,
default to fetching the archives. As builds from here on out should have
the proper untarred artifacts available, we should be able to stop
producing archives for system-ota-test builders after this change, and
the only dependence on archives should be for old builds which have
already produced those archives.

The new workflow is to download the `packages` directory from the
artifacts bucket, and download the blobs from `all_blobs.json`, and
create a package repository from that. Then download the bootserver and
images dir which contains the image manifest. The pave scripts will be
replaced by calling the bootserver tool with "--images images.json
--mode <pave or pave-zedboot> --<extra args>" within the image dir.

This change requires the artifacts tool to include fxr/448476.

Test: tested in CQ

Bug: 52021
Change-Id: I473868df0771fc0bd6294078d17d9184da179d54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438519
Commit-Queue: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>