[component_manager] Actions use SharedFuture
Actions represent an asynchronous operation on a component or realm
(such as stop, start, and delete). Actions are idempotent, in the sense
that two matching actions scheduled at the same time will only be
executed once.

Currently, this is implemented as a custom future. Change the
implementation to use SharedFuture, which is simpler.

A side effect of this change resulted in task ordering differences. This
caused some tests to fail which were inadvertently dependending on the
previous order (for example, assuming that when bind() returns, the
runner has actually started the component). Fix those tests to be
agnostic to the order.

Change-Id: Id96ea9a33063349198106e2ddb6b81d035fd11de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450182
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>