[fbl][docs] : Add fbl:: container docs.
Get started on writing a top level User's Guide for using the fbl
intrusive containers.

The idea here is to provide a conversational tone in an example heavy
doc which shows what the containers can do, and the various ways to do
the most common things.  Somewhere between a FAQ and a cookbook.

A comprehensive reference section is something which would be great to
have as well, but this is a start at least.

Bugs: 49748
Change-Id: I27b4e90dfa59b8f58a6d5789a08fcfb4499cd203
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389453
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>