[omaha-client]  Add timeout to HttpRequest::request()
This adds timeout-signalling support to HttpRequest, as Hyper itself
does not signal TCP read timeouts (which TCP doesn't guarantee the
detection of, as they are client-side decisions that the sender hasn't
sent expected data via the transport in an expected amount of time).

This changes the HttpRequest to return a collected response body as
a Vec of bytes, instead of a streaming Body, so that the
implementation can account for all timeouts in the reading of the
response from the server.

This adds a Timeout variant to the HttpRequest::ErrorKind struct, so
that the implementation of HttpRequest can flag that a Timeout
occurred, and so that the StateMachine can be tested for proper
behavior in those cases.

It adds a constructor method for implementations to use to signal
that a timeout has occured, and the fuchsia hyper implementation
in the omaha service binary uses that to signal timeouts when it
can't receive a response in a timely fashion.

Bug: 62668
Test: fx test fuchsia-hyper-test-support-tests
Test: fx test omaha-client-service-bin-tests
Test: fx test omaha-client-lib-tests
Change-Id: Ida88bf3fe66ebf99e661f9bf4aff86d6cf1c78d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445295
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>