[fidl][build] Isolate FIDL libraries in the same directory
This changes the FIDL build rules to provide greater separation between
FIDL libraries declared in the same BUILD.gn. Previously files and
directories were generated based on the location of the BUILD.gn in the
source tree and the FIDL library name. With this change the GN target
name is also always included.

This change has two effects:

First, for C++ every FIDL library declared in a directory will have a
different include path. This means that if a source_set has headers that
refer to a FIDL library those libraries must be in the public_deps, so
that when those headers are used outside of the source_set the compiler
will be able to find them. I had to update the netemul runner build
rules to take this into account.

Second, it's now possible to have two FIDL libraries with the same name
declared in the same BUILD.gn. Generally it's undesirable to have the
same FIDL library declared multiple times in the build but it's very
useful for testing.

Change-Id: I928de86e958b41b69dbc07932b2b56567eb25888
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442543
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>