[fs] Fix flaky test.
The problem here was to do with the fact that we do not have precise
control over ordering and it was possible for one of the Flush calls in
FlushPending to be slow and come after a flush from sync, which would
result in the test seeing an extra flush.  The fix is to:

 + Change FlushPending so that it doesn't actually Flush if there is no
   pending work.

 + Remove the Sync from the test, which was wrong anyway; the test is
   supposed to make sure it works without Sync.

In addition to the above, I have taken the opportunity to remove
barrier_ which was superfluous; journal_sequencer_ can be used instead.

Fixed: 63379
Change-Id: I613ff4ebc1bf769bad9d354a521cf462884d3647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449919
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Testability-Review: Manali Bhutiyani <manalib@google.com>
Commit-Queue: Chris Suter <csuter@google.com>