[components][logs] Routing log end-to-end test
Create a test topology where a component tries to access a
capability it does not have access to. In the test topology we
run Archivist and a reader component in the topology checks to
verify Archivists logs contain an error message for the routing
failure which looks about as expected and is attributed to the
requesting component instead of component manager.

Fixed: 63000

Change-Id: Iad54e8b87faa362aa1dd9801017cd7bb9c32184b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447075
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>