[build] Specify path to vendored python binary in exec_script
This updates the exec_script() invocations in this repository to
specify the full path to the vendored python binary to use instead of
relying on the PATH or global script_executable to find it. This also
means that the path to the script, along with any other input files,
needs to be passed to exec_script()'s file_dependencies parameter so
that the generated ninja file re-generates when these inputs change.

This also adds the -S flag to speed up Python initialization for these
scripts.

Bug: 63458, 62771
Change-Id: I7aa3e47f06cb301b39617f847b98b13f22f0eb78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446534
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>