[fidlc] Prevent buffer over-read in lexer
This CL fixes a bug where the lexer would advance past the end of a file
into adjacent memory. The Lex() function now asserts that it is not
called again after returning EOF. Satisfying this required changing some
of the recoverable parsing routines to be EOF-aware.

Here is the FIDL that triggered the bug:

   struct Foo {
      string
   };

It occurred like this:

1. Parse 1st field type "string".
2. Parse 1st field name "}". ERROR, expected an identifier.
3. Recover until end of field. Consume the ";".
4. Parse 2nd field type <EOF>. ERROR, expected an identifier.
5. Recover until end of field. Attempt to lex past end of buffer!

With this fix, step 5 returns early because EOF was already reached.

Note: It might seem more desirable to avoid interpreting the "}" as a
type name, and instead recognize it as ending the struct scope. I tried
doing this, but it required checking previous_token_ after consuming it
in a completely different part of the code, which feels hacky. There
might be a better way, but for now at least fidlc does not crash.

Test: fx test fidl-compiler
Fixed: 52698
Change-Id: Ie2bb69e104fce67f604bae6ed24ad1c0ef12c36f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451820
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>