[omaha client] correct calculation of update_check_response_time metric
The update_check_response_time metric is supposed to represent the
reponse time of making a single request to omaha (both success and
failure).  It was not being captured in the failure case, and in the
success case, it was the total time spent waiting between retries,
as well as the total time spent waiting for a failed attempt to occur.

This switches it to a per-attempt metric, also capturing the success
or failure of the result (the metric already had a dimension for this).

Bug: 60557
Bug: 60447
Bug: 60448
Change-Id: I390edd716b6f2721fd49c10a82252b9db805e5f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447298
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>