[kernel][x86][lbr] Add LBR k commands
Now that the global CpuidIo cache in the kernel is enabled, we can
properly use arch::LbrStack. As a proof-of-concept, we add an `lbr`
subcommand to `k cpu` that can enable, disable, and dump the LBR stacks
associated with particular CPUs.

Test: the k commands work on my NUC and return "not supported" in QEMU.
Change-Id: I26e6ad55accc6f068de1d35b95000b5aa2411d70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446380
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>