[scenic/escher] Injectable GpuAllocator.
Escher can now optionally be provided with an alternate allocator
as a constructor argument.  If none is provided, then a default
allocator will be instantiated.

The goal is to allow Escher clients to know the concrete type of the
the allocator, which is otherwise opaque (i.e. users of the default
allocators know it is a GpuAllocator, but not which subclass).  Since
the client may want to continue to reference the allocator, Escher
now holds the allocator with a shared_ptr instead of a unique_ptr.

Change-Id: I8d63a04de7e0ea86a4ee40b923b32e8bf52a24c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/449718
Reviewed-by: Chris Fontas <cfontas@google.com>
Testability-Review: Chris Fontas <cfontas@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>