[zircon_benchmarks] Clone VMOs wth multiple mappings
This extends the CloneMap VMO benchmarks to map the VMO using multiple
mappings, instead of a single mapping. The purpose of this is to
quantify the hidden impacts of VMO operations when multiple contiguous
mappings are used versus a single mapping, which is otherwise
functionally equivalent.

Bug: 42207

Change-Id: I7dc386e7a3686e529b7cae0d34dbea0a63aa5bfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448575
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Testability-Review: Mark Seaborn <mseaborn@google.com>