[fvm] Fix sparse image generation
Previously, when adding an empty minfs partition, there was some special
logic in sparse.cc which attempted to write out the empty partition
data.

There was a bug in this code which resulted in writing *all of the
slices*, rather than just the *extent* of data. This violates the sparse
image format, which expects only the sum of the extent_lengths to be
written out (the slices are virtual and are zero-extended when writing
out the sparse image as a real FVM image).

This bug went unnoticed because we always stick the corrupt minfs
partition at the end of the archive. Adding another partition after this
will result in the partition being miswritten, since we'd still be
reading from the empty minfs partition when we should be reading from
the first extent of the next partition.

Fix this by removing the special case logic for the empty minfs
partition; it is handled correctly by code below.

Change-Id: Id098a414c79a6adefda42b407bdabc7928bf03d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451975
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>