[fidlc] Support handle parametrization in aliases
Fixes handling of handle parametrization when used from a type alias.

We are purposefully not supporting partial type aliases since these are
being removed right now, as part of the implementation of FTP-052.

Interestingly, the `fuchsia.io2` library relied on this feature

    using Token = zx.handle:EVENT;

and the corrected handling in the compiler requires a source change. No
other FIDL file uses this sort of aliasing, and this change can land as
is without requiring an LSC.

Bug: 64629
Bug: 7807
Test: fx ninja -C out/default host_x64/fidl-compiler && \
      $FUCHSIA_DIR/out/default/host_x64/fidl-compiler --gtest_filter 'TypeAliasTests.*'
Change-Id: I0ae9567314596525a0dd346a80e5e0df7b85cfb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/310005
Commit-Queue: Pascal Perez <pascallouis@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>