[weave] Reduce the severity of logging.
Whenever a socket is closed, the corresponding fd waiter is cleared.
If no fd waiter is associated with the socket fd, it prints an error.
This is not a fatal error. So reducing the severity of the log.

Bug: 64405
Test: Verified locally that logs are now at proper severity.
Change-Id: Id9645f5b4b1b42d46aabd71f31e6bbd581e032b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451514
Commit-Queue: Sundarajan Srinivasan <sundarsrini@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>