[fx][target] Order addresses by time. ULA support.
Adds ULA support when getting an SSH-able address for the target.

This takes target addresses and orders them by timestamp. This is to
make it possible for a target to maintain connectability upon several
reboots (as the target's address is no longer derived from its mac).

This also removes the TargetAddrFetcher trait as it was not being used
(just for some minor cleanup).

Change-Id: I419b8a551baa853cc48c5d722b2faf74a6b1fc3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451521
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>