[blobfs] Set up client for external decompressor.
A thread safe client object that can be reused and a simpler wrapper for
the decompressor class that takes the client object. The client object
should attempt to set up the connection and fifo upon creation and upon
any loss of connectivity should recreate the connection on the next
attempted use.

Adds small integration test for the client with the component.

Change-Id: Ie0cb4be0699737fb2c4bf58ecd206ec75bef8d18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446259
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>