[lowpan][lowpan-spinel-driver] Fix network name always being "OpenThread"
Internally, lowpan-spinel-driver (and the lowpan API in general) treats
the network name as an opaque blob of bytes. However, Spinel expects
a UTF8 string for the network name property. Since UTF8 strings are
explicitly zero-terminated in Spinel, this was causing OpenThread to
reject the name.

This error took some time took some time to track down because
OpenThread [was not returning an error in this case][1]: it would
simply accept the input and not change the value.

We now convert the opaque blob to a real rust `String` and send that
instead of the opaque blob. This does two things: it verifies that
the bytes are valid UTF8 and it also ensures that the value is
zero terminated when sent over the wire.

The updated unit test has been verified to fail in the old case and
now passes with the new changes.

[1]: https://github.com/openthread/openthread/issues/5829

Test: fx test lowpan-tests
Change-Id: Ie63b3ba8f26d8dcb1ffac417c41a7b3be5cb150f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452934
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Testability-Review: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>