[weavestack] Shutdown on disconnect from netstack
When the FIDL connection to netstack is dropped, attempting to act on
sockets will return EPIPE (the underlying zx_channel is lost). In this
state, weavestack will continue to fail as it attempts to recover the
socket and will need to reconnect to the components it relies on. To aid
this, when the FIDL loss is detected, a task is now posted to the
application to request a shutdown of the main run loop.

A subsequent restart of the stack will have another chance to bind to
netstack and continue. This additionally prevents excess log spam from
weavestack when the system is in this state.

Bug: 64223
Test: fx test weavestack-unittests
Test: fx test weavestack-adaptation-unittests
Change-Id: Ib137f0de6e2356f41a1615d707dd6bb627465921
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451530
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>
Fuchsia-Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>