[bt][rfcomm] Clean up disconnection logic
Updates the clean up logic in the Session when a peer disconnects.

Context:
The RFCOMM Server maintains a map of Sessions indexed by the ID
of the connected peer. This map is updated lazily, meaning
stale Sessions will still exist in the map (even if the peer has already
disconnected). This is not an issue as entries are cleared and
reinitialized when the peer reconnects. However, multiplexed RFCOMM
channels will still remain opened (and clients none the wiser about peer
disconnection). Furthermore, component inspection will also reveal these
stale Sessions and can be misleading to debug.

This was not an issue pre- fxr/438439 because the SessionInner was not
shared, so when the peer disconnects the inner would be dropped (and
therefore cleaned up appropriately).

In order to maintain the proper RAII mechanics, this CL updates Session
to provide a termination signal, which can be polled by clients. The
RFCOMM Server uses this signal to eagerly evict the Session.

Test: Updated bt-rfcomm-tests
Change-Id: Id104c85fca9a15813783e1af627d17d2a2226004
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/451378
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>