[mexec] Introduce //src/bringup/lib/mexec
This change introduces //src/bringup/lib/mexec, which is planned to be
the new site for userspace mexec logic. The `mexec_zbi_test()` template
will be moved to a testing/ subdirectory in a follow-up.

There are currently two places that benefit from this shared code, both
of which are updated here to make use of it: netsvc and the mexec ZBI
test chain-loading logic.

With this change we deprecate the fuchsia.kernel.MexecBroker protocol,
which no longer has any users (it will be removed in a follow-up). The
only benefit originally to the introduction of a FIDL service seems to
have been (a) abstracting the propagation of the root resource and (b)
language agnosticism. Regarding (a), the plan of record is now to
propogate an "mexec" resource down to any site that wishes to mexec,
and so there isn't a priori a need to propagate a second service down
to make use of that (a library call can do that with far less
overhead); regarding (b), both mexec callsites are written in C++ and
we do not wish to proliferate a convenient means to mexec anyway.

Test: the mexec ZBI tests continue to run successfully (and both the
x64 + arm devices on my desk - as well as those in CQ - boot happily).

Bug: 57244
Change-Id: Ief5f9004e484146648a10435b6683cb7e05964e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448115
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>